+++ crates/stackable-telemetry/src/tracing.rs
@@ -6,12 +6,10 @@
 //!
 //! To get started, see [`Tracing`].
 
-use opentelemetry::KeyValue;
+use opentelemetry::{trace::TracerProvider, KeyValue};
 use opentelemetry_appender_tracing::layer::OpenTelemetryTracingBridge;
 use opentelemetry_sdk::{
-    logs::{self, LoggerProvider},
-    propagation::TraceContextPropagator,
-    trace, Resource,
+    logs::LoggerProvider, propagation::TraceContextPropagator, trace::Config, Resource,
 };
 use opentelemetry_semantic_conventions::resource;
 use snafu::{ResultExt as _, Snafu};
@@ -161,15 +159,15 @@ impl Tracing {
             .add_directive("h2=off".parse().expect("invalid directive"));
 
             let log_exporter = opentelemetry_otlp::new_exporter().tonic();
-            let otel_log =
-                opentelemetry_otlp::new_pipeline()
-                    .logging()
-                    .with_exporter(log_exporter)
-                    .with_log_config(logs::config().with_resource(Resource::new(vec![
-                        KeyValue::new(resource::SERVICE_NAME, self.service_name),
-                    ])))
-                    .install_batch(opentelemetry_sdk::runtime::Tokio)
-                    .context(InstallOtelLogExporterSnafu)?;
+            let otel_log = opentelemetry_otlp::new_pipeline()
+                .logging()
+                .with_exporter(log_exporter)
+                .with_resource(Resource::new(vec![KeyValue::new(
+                    resource::SERVICE_NAME,
+                    self.service_name,
+                )]))
+                .install_batch(opentelemetry_sdk::runtime::Tokio)
+                .context(InstallOtelLogExporterSnafu)?;
 
             // Convert `tracing::Event` to OpenTelemetry logs
             layers.push(
@@ -192,11 +190,12 @@ impl Tracing {
             let otel_tracer = opentelemetry_otlp::new_pipeline()
                 .tracing()
                 .with_exporter(trace_exporter)
-                .with_trace_config(trace::config().with_resource(Resource::new(vec![
+                .with_trace_config(Config::default().with_resource(Resource::new(vec![
                     KeyValue::new(resource::SERVICE_NAME, self.service_name),
                 ])))
                 .install_batch(opentelemetry_sdk::runtime::Tokio)
-                .context(InstallOtelTraceExporterSnafu)?;
+                .context(InstallOtelTraceExporterSnafu)?
+                .tracer(self.service_name);
 
             layers.push(
                 tracing_opentelemetry::layer()
@@ -207,7 +206,7 @@ impl Tracing {
 
             opentelemetry::global::set_text_map_propagator(
                 // NOTE (@NickLarsenNZ): There are various propagators. Eg: TraceContextPropagator
-                // standardises HTTP headers to propagate trace-id, parent-id, etc... while the
+                // standardizes HTTP headers to propagate trace-id, parent-id, etc... while the
                 // BaggagePropagator sets a "baggage" header with the value being key=value pairs. There
                 // are other kinds too. There is also B3 and Jaeger, and some legacy stuff like OT Trace
                 // and OpenCensus.